+++ src/main/java/com/example/DoroServer/domain/lectureContent/service/LectureContentService.java
@@ -9,15 +9,12 @@ import com.example.DoroServer.domain.lectureContent.repository.LectureContentRep
 import com.example.DoroServer.domain.lectureContentImage.entity.LectureContentImage;
 import com.example.DoroServer.global.exception.BaseException;
 import com.example.DoroServer.global.exception.Code;
-import com.example.DoroServer.global.s3.AwsS3ServiceImpl;
 
-import java.util.Arrays;
-import java.util.List;
 import java.util.stream.Collectors;
+import java.util.List;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -25,10 +22,10 @@ import lombok.RequiredArgsConstructor;
 @Transactional
 public class LectureContentService {
 
-    private final AwsS3ServiceImpl awsS3Service;
     private final LectureContentRepository lectureContentRepository;
     private final LectureContentMapper lectureContentMapper;
     private final ModelMapper modelMapper;
+    private final ContentImageS3ServiceImpl contentImageS3Service;
 
     /* 서비스 코드 */
 
@@ -39,13 +36,14 @@ public class LectureContentService {
     }
 
     public LectureContentRes createLectureContent(CreateLectureContentReq lectureContentReq) {
-        validateDtoFiles(lectureContentReq.getFiles());
-
         LectureContent lectureContent = modelMapper.map(lectureContentReq, LectureContent.class);
 
-        List<String> uploadedUrls = uploadFilesAndGetUrls(lectureContentReq.getFiles());
+        // lectureContentReq의 files를 S3에 업로드 후, 업로드된 파일들의 URL을 LectureContentImage로 변환
+        List<String> uploadedUrls = contentImageS3Service.uploadS3Images(lectureContentReq.getFiles());
+        List<LectureContentImage> lectureContentImages = uploadedUrls.stream()
+                .map(LectureContentImage::new)
+                .collect(Collectors.toList());
 
-        List<LectureContentImage> lectureContentImages = createLectureContentImagesWith(uploadedUrls);
         lectureContent.getImages().addAll(lectureContentImages);
         LectureContent savedLectureContent = lectureContentRepository.save(lectureContent);
 
@@ -55,7 +53,7 @@ public class LectureContentService {
     public LectureContentRes updateLectureContent(Long id, UpdateLectureContentReq updateLectureContentReq) {
         LectureContent lectureContent = findLectureContentById(id);
 
-        modelMapper.map(updateLectureContentReq, lectureContent); // 필드있고없고차이 MapStruct와 비교
+        modelMapper.map(updateLectureContentReq, lectureContent);
 
         LectureContent updatedLectureContent = lectureContentRepository.save(lectureContent);
 
@@ -65,56 +63,16 @@ public class LectureContentService {
     public void deleteLectureContent(Long id) {
         LectureContent lectureContent = findLectureContentById(id);
 
-        lectureContent.getImages().forEach(this::deleteS3Image);
+        lectureContent.getImages().forEach(contentImageS3Service::deleteS3Image);
 
         lectureContentRepository.delete(lectureContent);
     }
 
     /* 서비스 코드에서 사용되는 메서드 */
 
-    void validateDtoFiles(MultipartFile[] files) {
-        if (files == null) {
-            // Multipart가 한 개라도 배열에 넣어서 요청해야함
-            throw new BaseException(Code.JSON_SYNTAX_ERROR);
-        }
-        if (files.length > 100) {
-            throw new BaseException(Code.LECTURE_CONTENT_IMAGE_COUNT_OVER);
-        }
-        for (MultipartFile file : files) {
-            if (file.getSize() > 10485760) { // 10MB
-                throw new BaseException(Code.LECTURE_CONTENT_IMAGE_SIZE_OVER);
-            }
-        }
-    }
-
     LectureContent findLectureContentById(Long id) {
         return lectureContentRepository.findById(id)
                 .orElseThrow(() -> new BaseException(Code.LECTURE_CONTENT_NOT_FOUND));
-    }
-
-    List<LectureContentImage> createLectureContentImagesWith(List<String> uploadedUrls) {
-        // 업로드된 파일들의 URL로 LectureContentImage를 만들어 리스트로 반환
-        return uploadedUrls.stream()
-                .map(url -> LectureContentImage.builder().url(url).build())
-                .collect(Collectors.toList());
-    }
-
-    List<String> uploadFilesAndGetUrls(MultipartFile[] files) {
-        // S3 BUCKET/lecture-content에 파일 업로드 후, 업로드된 파일들의 URL 반환
-        return Arrays.stream(files)
-                .map(file -> awsS3Service.upload(file, "lecture-content"))
-                .collect(Collectors.toList());
-    }
-
-    void deleteS3Image(LectureContentImage lectureContentImage) {
-        String uploadedUrl = lectureContentImage.getUrl();
-        String fileName = getFileNameFrom(uploadedUrl);
-        awsS3Service.deleteImage(fileName);
-    }
-
-    String getFileNameFrom(String url) {
-        // UUID(36자) + .jpg(4자) = 파일명(40자)
-        return url.substring(40);
     }
 
 }
